+++ Samples~/Read output asynchronously/AsyncReadbackCompute.cs
@@ -8,28 +8,22 @@ public class AsyncReadbackCompute : MonoBehaviour
 
     Tensor m_Input;
     IWorker m_Engine;
-    TensorFloat m_OutputTensor;
 
-    void ReadbackCallback(bool completed)
+    async void OnEnable()
     {
-        // Put the downloaded tensor data into a readable tensor before indexing.
-        m_OutputTensor.MakeReadable();
-        Debug.Assert(m_OutputTensor[0] == 42);
-        Debug.Log($"Output tensor value {m_OutputTensor[0]}");
-    }
-
-    void OnEnable()
-    {
-        // Everything that can be statically assigned is setup during Start to avoid memory churn.
         var model = ModelLoader.Load(modelAsset);
         m_Input = new TensorFloat(new TensorShape(1, 1), new[] { 43.0f });
         m_Engine = WorkerFactory.CreateWorker(BackendType.GPUCompute, model);
-        m_Engine.Execute(m_Input);
 
-        // Peek the value from Sentis, without taking ownership of the Tensor (see PeekOutput docs for details).
-        m_OutputTensor = m_Engine.PeekOutput() as TensorFloat;
-        m_OutputTensor.AsyncReadbackRequest(ReadbackCallback);
-        Debug.Log($"This code is called before the callback.");
+        m_Engine.Execute(m_Input);
+        var outputTensor = m_Engine.PeekOutput() as TensorFloat;
+
+        await outputTensor.ReadbackRequestAsync();
+
+        // Put the downloaded tensor data into a readable tensor before indexing.
+        outputTensor.MakeReadable();
+        Debug.Assert(outputTensor[0] == 42);
+        Debug.Log($"Output tensor value {outputTensor[0]}");
     }
 
     void OnDisable()